+++ src/application/listeners.js
@@ -6,13 +6,13 @@ const sprintf = require('sprintf-js').sprintf;
 const {
     get_command_by_keyword,
     get_command_by_command,
-    TRANSLATIONS,
-} = require.main.require('./parts/parameters.js');
+    get_translation,
+} = require.main.require('./setup/config.js');
 const {
     remove_message,
     delay_remove_reply,
-} = require.main.require('./parts/operations.js');
-const { get_main_menu } = require.main.require('./parts/menus.js');
+} = require.main.require('./application/operations.js');
+const { get_main_menu } = require.main.require('./application/menus.js');
 
 /****************************************************************
  * METHODS - LISTENERS
@@ -27,7 +27,7 @@ const listener_on_callback_query = async (bot, msg, {debug, timeout}) => {
     const keyword = msg.data;
     const commands = get_command_by_keyword(keyword);
     const { group } = commands[0] || '';
-    const message = TRANSLATIONS.value(lang, 'redirect-message');
+    const message = get_translation(lang, 'redirect-message');
     const responseText = `${message}: ${group}`;
     const options = {};
     if (debug) console.debug(`:callback_query: id=${chatId}; lang=${lang}; keyword=${keyword}, group=${group}.`);
@@ -48,7 +48,7 @@ const listener_on_message = async (bot, msg, {debug, timeout, timeout_menu}) => 
 
     // special treatment if new chat member:
     if ('new_chat_members' in msg) {
-        const responseText = sprintf(TRANSLATIONS.value(lang, 'welcome-message'), username);
+        const responseText = sprintf(get_translation(lang, 'welcome-message'), username);
         const options = get_main_menu(lang);
 
         // Trigger response, then delete response after delay:
@@ -65,7 +65,7 @@ const listener_on_message = async (bot, msg, {debug, timeout, timeout_menu}) => 
     const commands = get_command_by_command(command);
     // Response to `/hello` - command only available in debug mode:
     if (debug && (command == '/hello')) {
-        const responseText = sprintf(TRANSLATIONS.value(lang, 'welcome-message'), username);
+        const responseText = sprintf(get_translation(lang, 'welcome-message'), username);
         const options = {};
 
         // Delete message, trigger response, then delete response after delay:
@@ -74,7 +74,7 @@ const listener_on_message = async (bot, msg, {debug, timeout, timeout_menu}) => 
         delay_remove_reply(bot, timeout, reply);
     // Response to `/start`
     } else if (command == '/start') {
-        const responseText = TRANSLATIONS.value(lang, 'start-message');
+        const responseText = get_translation(lang, 'start-message');
         const options = get_main_menu(lang);
 
         // Trigger response, then delete response after delay:
@@ -82,7 +82,7 @@ const listener_on_message = async (bot, msg, {debug, timeout, timeout_menu}) => 
         delay_remove_reply(bot, timeout_menu, reply);
     // Response to `/help`
     } else if (command == '/help') {
-        const responseText = TRANSLATIONS.value(lang, 'help-message');
+        const responseText = get_translation(lang, 'help-message');
         const options = get_main_menu(lang);
 
         // Delete message, trigger response, then delete response after delay:
@@ -91,7 +91,7 @@ const listener_on_message = async (bot, msg, {debug, timeout, timeout_menu}) => 
         delay_remove_reply(bot, timeout_menu, reply);
     // Response to command from list:
     } else if (commands.length > 0) {
-        const message = TRANSLATIONS.value(lang, 'redirect-message');
+        const message = get_translation(lang, 'redirect-message');
         const options = {};
         const { group } = commands[0] || '';
         const responseText = `${message}: ${group}`;
