+++ src/setup/config.yaml
@@ -14,6 +14,7 @@ languages:
   - pl
   - ru
   - uk
+default-language: en
 # ------------------------------------------------------------------------------------------------
 # COMMANDS
 # Structure:
